+++ common/src/main/java/com/gitlab/controllers/api/rest/PersonalAddressRestApi.java
@@ -1,19 +1,27 @@
-package com.gitlab.controller.api;
+package com.gitlab.controllers.api.rest;
 
 import com.gitlab.dto.PersonalAddressDto;
-import io.swagger.annotations.*;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
+import io.swagger.annotations.ApiResponse;
+import io.swagger.annotations.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import javax.validation.Valid;
 import java.util.List;
 
 @Api(tags = "Personal Address REST")
 @Tag(name = "Personal Address REST", description = "API for personal address")
-@RequestMapping("/api/personal_address")
 public interface PersonalAddressRestApi {
-    @GetMapping
+    @GetMapping("/api/personal_address")
     @ApiOperation(value = "Get all Personal Addresses")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Personal Addresses found"),
@@ -21,16 +29,16 @@ public interface PersonalAddressRestApi {
     )
     ResponseEntity<List<PersonalAddressDto>> getAll();
 
-    @GetMapping("/{id}")
+    @GetMapping("/api/personal_address/{id}")
     @ApiOperation(value = "Get Personal Address by id")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Personal Address found"),
             @ApiResponse(code = 404, message = "Personal Address not found")}
     )
     ResponseEntity<PersonalAddressDto> get(@ApiParam(name = "id", value = "PersonalAddress.id")
-                                           @PathVariable Long id);
+                                           @PathVariable (value = "id") Long id);
 
-    @PostMapping
+    @PostMapping("/api/personal_address")
     @ApiOperation(value = "Create Personal Address")
     @ApiResponses(value = {
             @ApiResponse(code = 201, message = "Personal Address created"),
@@ -39,23 +47,23 @@ public interface PersonalAddressRestApi {
     ResponseEntity<PersonalAddressDto> create(@ApiParam(name = "personalAddress", value = "PersonalAddressDto")
                                               @Valid @RequestBody PersonalAddressDto personalAddressDto);
 
-    @PatchMapping("/{id}")
+    @PatchMapping("/api/personal_address/{id}")
     @ApiOperation(value = "Update Personal Address")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Personal Address updated"),
             @ApiResponse(code = 400, message = "Personal Address not updated")}
     )
     ResponseEntity<PersonalAddressDto> update(@ApiParam(name = "id", value = "PersonalAddress.id")
-                                              @PathVariable Long id,
+                                              @PathVariable (value = "id") Long id,
                                               @ApiParam(name = "personalAddress", value = "PersonalAddressDto")
                                               @Valid @RequestBody PersonalAddressDto personalAddressDto);
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/api/personal_address/{id}")
     @ApiOperation(value = "Delete Personal Address by id")
     @ApiResponses(value = {
             @ApiResponse(code = 200, message = "Personal Address deleted"),
             @ApiResponse(code = 404, message = "Personal Address not found")}
     )
     ResponseEntity<Void> delete(@ApiParam(name = "id", value = "PersonalAddress.id")
-                                @PathVariable Long id);
+                                @PathVariable (value = "id") Long id);
 }