+++ gazon-service/src/test/java/com/gitlab/mapper/UserMapperTest.java
@@ -2,6 +2,9 @@ package com.gitlab.mapper;
 
 import com.gitlab.controller.AbstractIntegrationTest;
 import com.gitlab.dto.*;
+import com.gitlab.enums.Citizenship;
+import com.gitlab.enums.EntityStatus;
+import com.gitlab.enums.Gender;
 import com.gitlab.model.*;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,7 +40,7 @@ class UserMapperTest extends AbstractIntegrationTest {
 
         Passport passport = new Passport(
                 1L,
-                Passport.Citizenship.RUSSIA,
+                Citizenship.RUSSIA,
                 "user",
                 "user",
                 "patronym",
@@ -55,13 +58,14 @@ class UserMapperTest extends AbstractIntegrationTest {
                 "user",
                 "user",
                 LocalDate.now(),
-                User.Gender.MALE,
+                Gender.MALE,
                 "89007777777",
                 passport,
                 LocalDate.now(),
                 bankCardSet,
                 personalAddresses,
-                roleSet);
+                roleSet,
+                EntityStatus.ACTIVE);
 
         UserDto actualResult = mapper.toDto(user);
 
@@ -130,7 +134,7 @@ class UserMapperTest extends AbstractIntegrationTest {
 
         PassportDto passport = new PassportDto(
                 1L,
-                Passport.Citizenship.RUSSIA,
+                Citizenship.RUSSIA,
                 "user",
                 "user",
                 "patronym",
@@ -149,7 +153,7 @@ class UserMapperTest extends AbstractIntegrationTest {
                 "user",
                 "user",
                 LocalDate.now(),
-                User.Gender.MALE,
+                Gender.MALE,
                 "89007777777",
                 passport,
                 personalAddress,