+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 - #77 - oauth через гугл
 - #66 - Front CRUD for Product
 - #72 - Front Crud for WorkingSchedule
+- #105 - Status model for User
+- #107 - Search correction
 
 ## [1.1.0] - 8.09.2023
 - #5 - added user entity